refact: Separate package for shared utilities

+++ runner/utils/scheduler.go
@@ -2,7 +2,7 @@ package utils
 
 type RADSDecision struct {
 	AvailableSlots int
-	IdleSlots      int
+	IdleSlots      float64
 	UsedSlots      int
 	Status         string
 }
